+++ morpheus/framework/resources/networkrouterfirewallrulegroup/delete.go
@@ -1,0 +1,59 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
+package networkrouterfirewallrulegroup
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+
+	"github.com/HPE/terraform-provider-hpe/morpheus/utils/errfmt"
+)
+
+const deleteOperation = "delete network router firewall rule group resource"
+
+func (r *Resource) Delete(
+	ctx context.Context,
+	req resource.DeleteRequest,
+	resp *resource.DeleteResponse,
+) {
+	var state NetworkRouterFirewallRuleGroupModel
+
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	client, err := r.NewClient(ctx)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			deleteOperation,
+			"failed to create client: "+err.Error(),
+		)
+
+		return
+	}
+
+	id := state.Id.ValueInt64()
+	routerID := state.RouterId.ValueInt64()
+
+	tflog.Debug(ctx, fmt.Sprintf("Deleting firewall rule group %d on router %d", id, routerID))
+
+	_, hresp, err := client.NetworksAPI.
+		DeleteNetworkRouterFirewallRuleGroup(ctx, id, routerID).Execute()
+	if err != nil {
+		if hresp != nil && hresp.StatusCode == http.StatusNotFound {
+			tflog.Debug(ctx, fmt.Sprintf("Firewall rule group %d already deleted (404)", id))
+
+			return
+		}
+
+		resp.Diagnostics.AddError(
+			deleteOperation,
+			fmt.Sprintf("firewall rule group %d DELETE failed: %s", id, errfmt.ErrMsg(err, hresp)),
+		)
+	}
+}
