+++ morpheus/framework/resources/networkrouterfirewallrulegroup/resource_test.go
@@ -1,3 +1,5 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
 package networkrouterfirewallrulegroup_test
 
 import (
